SA-0MPPLELMX0025D5U: Convert ship agent to a pi-agent skill
Create skill/ship/ directory as a proper pi-agent skill with SKILL.md
following the pi-agent skill format. Move agent/ship.js and
agent/git-helpers.js into skill/ship/scripts/ with updated module
exports. Retain original agent/ files as thin re-export wrappers
for backward compatibility.

Update agent/ship.md to reference the skill rather than containing
inline guidance. Update test imports to point to the new skill
module locations.

All 62 JS tests pass (unit + integration + pre-push hook + critical).
Full pytest suite: 496 pass, 3 pre-existing failures unchanged.
