+++ skill/ship/__init__.py
@@ -1,0 +1,1 @@
+# Ship skill package
